requirements.txt
pytest~=7.3.2
django~=4.2.2
django-crispy-forms~=2.0
pytz~=2023.3